Fixed a few issues
Problem: When the user is already added to persistence, and user signs in for the first time the dummy values in persistence were not getting updated.
Solution: Updating all the user details when the user signs in for the first time. When we fill the strategy vs id on the user object for the first time.

Problem: OauthToken constructor would throw Error when values are null, but null are usually expected for OauthToken
Solution: Updated the constructor to accept nulls too.